% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc}
\alias{fastqc}
\title{Run FastQC Tool}
\usage{
fastqc(fq.dir = getwd(), qc.dir = NULL, threads = 4,
  fastqc.path = "~/bin/FastQC/fastqc")
}
\arguments{
\item{fq.dir}{path to the directory containing fastq files. Default is the
current working directory.}

\item{qc.dir}{path to the FastQC result directory. If NULL, a directory
named fastqc_results is created in the current working directory.}

\item{threads}{the number of threads to be used. Default is 4.}

\item{fastqc.path}{path to fastqc program}
}
\value{
Create a directory containing the reports
}
\description{
Run FastQC Tool
}
\examples{
\dontrun{
# Run FastQC: generates a QC directory
fastqc(fq.dir)
}
}
