% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc_install.R
\name{fastqc_install}
\alias{fastqc_install}
\title{Install FastQC Tool}
\usage{
fastqc_install(url, dest.dir = "~/bin")
}
\arguments{
\item{url}{url to download the latest version. If missing, the function will
try to install the latest version from
\href{http://www.bioinformatics.babraham.ac.uk/projects/download.html#fastqc}{http://www.bioinformatics.babraham.ac.uk/projects/download.html#fastqc}.}

\item{dest.dir}{destination directory to install the tool.}
}
\description{
Install the FastQC Tool. To be used only on Unix system.
}
