% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot_collection.R
\name{plot_gc_content_collection}
\alias{plot_gc_content_collection}
\title{Plot GC content of a collection of samples}
\usage{
plot_gc_content_collection(qc, ggtheme = theme_minimal(), ...)
}
\arguments{
\item{qc}{An object of class qc_read_collection}

\item{ggtheme}{A plotting themem}

\item{...}{Other}
}
\value{
A graph of mulitple lines each corresponds to the GC content from one sample.
}
\description{
A working example of a function to plot the GC content of multiple samples as multiple lines.
}
\examples{
# extract paths to the demo files
qc.dir <- system.file("fastqc_results", package = "fastqcr")
qc.files <- list.files(qc.dir, full.names = TRUE)

# read all modules in all files
qc <- qc_read_collection(qc.files, sample_names = paste('S', 1:5, sep = ''))

# plot GC content in all samples
plot_gc_content_collection(qc)

}
