% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unzip.R
\name{qc_unzip}
\alias{qc_unzip}
\title{Unzip Files in the FastQC Result Directory}
\usage{
qc_unzip(qc.dir = ".", rm.zip = TRUE)
}
\arguments{
\item{qc.dir}{Path to the FastQC result directory.}

\item{rm.zip}{logical. If TRUE, remove zipped files after extraction. Default
is TRUE.}
}
\description{
Unzip all files in the FastQC result directory. Default is the
  current working directory.
}
\examples{
\dontrun{
qc_unzip("FASTQC")
}
}
